+++ modules/stb_vorbis/resource_importer_ogg_vorbis.cpp
@@ -35,16 +35,13 @@
 #include "scene/resources/texture.h"
 
 String ResourceImporterOGGVorbis::get_importer_name() const {
-
 	return "ogg_vorbis";
 }
 
 String ResourceImporterOGGVorbis::get_visible_name() const {
-
 	return "OGGVorbis";
 }
 void ResourceImporterOGGVorbis::get_recognized_extensions(List<String> *p_extensions) const {
-
 	p_extensions->push_back("ogg");
 }
 
@@ -53,12 +50,10 @@ String ResourceImporterOGGVorbis::get_save_extension() const {
 }
 
 String ResourceImporterOGGVorbis::get_resource_type() const {
-
 	return "AudioStreamOGGVorbis";
 }
 
 bool ResourceImporterOGGVorbis::get_option_visibility(const String &p_option, const Map<StringName, Variant> &p_options) const {
-
 	return true;
 }
 
@@ -66,18 +61,15 @@ int ResourceImporterOGGVorbis::get_preset_count() const {
 	return 0;
 }
 String ResourceImporterOGGVorbis::get_preset_name(int p_idx) const {
-
 	return String();
 }
 
 void ResourceImporterOGGVorbis::get_import_options(List<ImportOption> *r_options, int p_preset) const {
-
 	r_options->push_back(ImportOption(PropertyInfo(Variant::BOOL, "loop"), true));
 	r_options->push_back(ImportOption(PropertyInfo(Variant::REAL, "loop_offset"), 0));
 }
 
 Error ResourceImporterOGGVorbis::import(const String &p_source_file, const String &p_save_path, const Map<StringName, Variant> &p_options, List<String> *r_platform_variants, List<String> *r_gen_files, Variant *r_metadata) {
-
 	bool loop = p_options["loop"];
 	float loop_offset = p_options["loop_offset"];
 
@@ -85,7 +77,7 @@ Error ResourceImporterOGGVorbis::import(const String &p_source_file, const Strin
 
 	ERR_FAIL_COND_V_MSG(!f, ERR_CANT_OPEN, "Cannot open file '" + p_source_file + "'.");
 
-	size_t len = f->get_len();
+	uint64_t len = f->get_len();
 
 	PoolVector<uint8_t> data;
 	data.resize(len);
